Display current game score on new game start
When after finishing a game the new one is started until the first
answer the score for already finished game is shown, which is pretty
awkward (answered no question, having 5 points of ten). What this commit
does is simply refreshes score display so users could see the score of
current game.